requirements.txt
py-clob-client>=0.18.0
web3>=7.6.1
python-dotenv>=1.0.1
python-telegram-bot>=21.6
fastapi>=0.115.0
uvicorn>=0.30.6
httpx>=0.27.2
pydantic>=2.9.2
annotated-doc>=0.0.2
# Note: bip-utils is optional (only needed for wallet generation)
# bip-utils==2.9.3

